Function to close mobile nav on link click
I may have the understanding wrong, but you don't want to add the category filter in the mobile nav if that is what you were thinking of. Instead you should have a category sort on the main part of the page, that can be passed down. It would work similar to the side nav. It would be a lot better if you built a context around these functions, state updates on page change, but context doesn't. Start researching that, it'll go far in the build